refactor(apps/earn): approve2 system for create
